# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  request.Offer:
    properties:
      description:
        maxLength: 50
        minLength: 6
        type: string
      discount_rate:
        maximum: 100
        minimum: 1
        type: integer
      end_date:
        type: string
      offer_name:
        type: string
      start_date:
        type: string
    required:
    - description
    - discount_rate
    - end_date
    - offer_name
    - start_date
    type: object
  request.OfferCategory:
    properties:
      category_id:
        type: integer
      offer_id:
        type: integer
    required:
    - category_id
    - offer_id
    type: object
  request.OfferProduct:
    properties:
      offer_id:
        type: integer
      product_id:
        type: integer
    required:
    - offer_id
    - product_id
    type: object
...
  request.UpdateCategoryOffer:
    properties:
      category_offer_id:
        type: integer
      offer_id:
        type: integer
    required:
    - category_offer_id
    - offer_id
    type: object
...
  request.UpdateProductOffer:
    properties:
      offer_id:
        type: integer
      product_offer_id:
        type: integer
    required:
    - offer_id
    - product_offer_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: okeke98@gmail.com
    name: For API Support
  description: 'Backend API built with Golang using Clean Code architecture. \nGithub:
    [https://github.com/okekefrancis112/ecommerce-gin].'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-commerce Application Backend API
paths:
...
  /admin/offers:
    get:
      description: API for admin to get all offers
      operationId: GetAllOffers
      parameters:
      - description: Page Number
        in: query
        name: page_number
        type: integer
      - description: Count
        in: query
        name: count
        type: integer
      responses:
        "200":
          description: "Successfully\tfound\tall\toffers"
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get all offers
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all offers (Admin)
      tags:
      - Admin Offers
    post:
      description: API for admin to add an offer (Admin)
      operationId: SaveOffer
      parameters:
      - description: input field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.Offer'
      responses:
        "200":
          description: Successfully offer added
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Offer already exist
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add offer (Admin)
      tags:
      - Admin Offers
  /admin/offers/{offer_id}:
    delete:
      description: API admin to remove an offer
      operationId: RemoveOffer
      parameters:
      - description: Offer ID
        in: path
        name: offer_id
        required: true
        type: integer
      responses:
        "200":
          description: successfully offer added
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove offer (Admin)
      tags:
      - Admin Offers
  /admin/offers/category:
    get:
      description: API for admin to get all category offers
      operationId: GetAllCategoryOffers
      parameters:
      - description: Page Number
        in: query
        name: page_number
        type: integer
      - description: Count
        in: query
        name: count
        type: integer
      responses:
        "200":
          description: successfully got all offer_category
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: failed to get offers_category
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all category offers (Admin)
      tags:
      - Admin Offers
    patch:
      description: API admin to change product offer to another offer
      operationId: ChangeCategoryOffer
      parameters:
      - description: input field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCategoryOffer'
      responses:
        "200":
          description: successfully offer replaced for category
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Change product offer (Admin)
      tags:
      - Admin Offers
    post:
      description: API for admin to add an offer category
      operationId: SaveCategoryOffer
      parameters:
      - description: input field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.OfferCategory'
      responses:
        "200":
          description: successfully offer added for category
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add category offer (Admin)
      tags:
      - Admin Offers
  /admin/offers/category/{offer_category_id}:
    delete:
      description: API admin to remove a offer from category
      operationId: RemoveCategoryOffer
      parameters:
      - description: Offer Category ID
        in: path
        name: offer_category_id
        required: true
        type: integer
      responses:
        "200":
          description: successfully offer added for category
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove category offer (Admin)
      tags:
      - Admin Offers
  /admin/offers/products:
    get:
      description: API for admin to get all product offers
      operationId: GetAllProductsOffers
      parameters:
      - description: Page Number
        in: query
        name: page_number
        type: integer
      - description: Count
        in: query
        name: count
        type: integer
      responses:
        "200":
          description: successfully got all offers_categories
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: failed to get offer_products
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all product offers (Admin)
      tags:
      - Admin Offers
    patch:
      description: API admin to change product offer to another offer
      operationId: ChangeProductOffer
      parameters:
      - description: input field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductOffer'
      responses:
        "200":
          description: Successfully offer changed for  given product offer
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Change product offer (Admin)
      tags:
      - Admin Offers
    post:
      description: API for admin to add an offer for product
      operationId: SaveProductOffer
      parameters:
      - description: input field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.OfferProduct'
      responses:
        "200":
          description: successfully offer added for product
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add product offer (Admin)
      tags:
      - Admin Offers
  /admin/offers/products/{offer_product_id}:
    delete:
      description: API admin to remove a offer from product
      operationId: RemoveProductOffer
      parameters:
      - description: offer_product_id
        in: path
        name: offer_product_id
        required: true
        type: integer
      responses:
        "200":
          description: Successfully offer removed from product
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: invalid input on params
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove product offer (Admin)
      tags:
      - Admin Offers
...
securityDefinitions:
  BearerAuth:
    description: 'Add prefix of Bearer before  token Ex: "Bearer token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"